+++ navigation/RELEASE_NOTES.md
@@ -4,3 +4,4 @@
 ##### Navigation
 * ANDDEP-312 Added navigation module
 * ANDDEP-1113 Fixed fragmentManager crash, on adding fragment when app is in background.
+* Fragment manager crash fix updated for Xiaomi on Android 9